requirements/base.txt
celery==3.1.23

Django==1.9.4
dj-database-url==0.4.0
django-model-utils==2.4
django-cors-headers==1.1.0
django-avatar==3.0.0
django-jinja==2.1.2
django-filter==0.13.0
django-crispy-forms==1.6.0
easy-thumbnails==2.3
python-social-auth==0.2.16


# sending emails
djmail==0.12.0.post1

# Django REST Framework
djangorestframework==3.3.3
git+https://github.com/silverlogic/drf-extra-fields.git@c5845d4de9873950d65bad3f487a6dc62f544e08

# runs the app server
gunicorn==19.4.5

# postgresql driver
psycopg2==2.6.1
